improve view json add support polling

+++ communityforum/views.py
@@ -98,27 +98,29 @@ def communitySeePost(request,pk):
 
 @login_required(login_url='centBaseLoginUser')
 def communityAddComment(request):
-    try:
-        data = json.loads(request.body)
-        postId = data['postId']
-        commentData = data['commentData']
-        tgtpost = ForumPost.objects.get(id=postId)
-    except:
-        return JsonResponse({'success':False, 'error':'invalid'})
-    
-    if str(commentData).strip() == "":
-        return JsonResponse({'success':False, 'error':'empty'})
-    
-    createdComment = ForumComment.objects.create(
-        user = request.user,
-        forumpost = tgtpost,
-        comment = commentData,
-    )
-    
-    ctxDict = {'createdcomment':commentData,'success': True}
-    messages.success(request, "Comment posted!")
-    return JsonResponse(ctxDict)
-    
+    if request.method == "POST":
+        try:
+            data = json.loads(request.body)
+            postId = data['postId']
+            commentData = data['commentData']
+            tgtpost = ForumPost.objects.get(id=postId)
+        except:
+            return JsonResponse({'success':False, 'error':'invalid'})
+        
+        if str(commentData).strip() == "":
+            return JsonResponse({'success':False, 'error':'empty'})
+        
+        createdComment = ForumComment.objects.create(
+            user = request.user,
+            forumpost = tgtpost,
+            comment = commentData,
+        )
+        
+        ctxDict = {'createdcomment':commentData,'success': True}
+        messages.success(request, "Comment posted!")
+        return JsonResponse(ctxDict)
+    else:
+        return HttpResponse('Invalid request')
 
 # @login_required(login_url='centBaseLoginUser')
 # def communityDeleteComment(request,pk):
@@ -137,17 +139,20 @@ def communityAddComment(request):
 
 @login_required(login_url='centBaseLoginUser')
 def communityDeleteComment(request):
-    try:
-        data = json.loads(request.body)
-        commentId = data['commentId']
-        commenttd = ForumComment.objects.get(id=commentId)
-    except:
-        # messages.warning(request, "comment not found")
-        return JsonResponse({"success":False})
+    if request.method == "POST":
+        try:
+            data = json.loads(request.body)
+            commentId = data['commentId']
+            commenttd = ForumComment.objects.get(id=commentId)
+        except:
+            # messages.warning(request, "comment not found")
+            return JsonResponse({"success":False})
 
-    if commenttd.user == request.user:
-        commenttd.delete()
-        messages.success(request,'The comment has been deleted.')
-        return JsonResponse({'message': 'The comment has been deleted', 'success':True})
+        if commenttd.user == request.user:
+            commenttd.delete()
+            messages.success(request,'The comment has been deleted.')
+            return JsonResponse({'message': 'The comment has been deleted', 'success':True})
+        else:
+            return HttpResponse('Unauthorized')
     else:
-        return HttpResponse('Unauthorized')
+        return HttpResponse('Invalid request')
